test/e2e/requirements.txt
pytest==7.4.0
psutil==5.9.5
pillow==10.0.0
opencv-python-headless==4.8.0.74
numpy~=1.25.1
pytesseract==0.3.10
atomacos==3.3.0; platform_system == "Darwin"
allure-pytest==2.13.2
testrail-api==1.12.0
pyperclip==1.8.2
pytest-rerunfailures==13.0
pytest-ignore-flaky==2.1.0
pytest-timeout==2.2.0
shortuuid==1.0.12
pluggy==1.2.0
